requirements.txt
Flask
gunicorn
numpy
pandas
scikit-learn
azure-storage-blob
